Support for graphics in the terminal.
Added a new shader program program to render textures in the grid.

Part of the new implementation uses GL_ARB_clear_texture and GL_ARB_copy_image
OpenGL extension, with a fallback in the hardware does not support them.

Added a parser for Sixel images.

memoffset crate is added to initialize vertex attributes.